support for lending market prices

+++ defitrack-rest/defitrack-api/src/main/java/io/defitrack/market/lending/vo/LendingPositionVO.kt
@@ -3,13 +3,14 @@ package io.defitrack.market.lending.vo
 import io.defitrack.network.NetworkVO
 import io.defitrack.protocol.ProtocolVO
 import io.defitrack.token.FungibleToken
+import java.math.BigDecimal
 import java.math.BigInteger
 
 data class LendingPositionVO(
     val id: String,
     val network: NetworkVO,
     val protocol: ProtocolVO,
-    val dollarValue: Double,
+    val dollarValue: BigDecimal,
     val rate: Double?,
     val name: String,
     val amountDecimal: Double,
